Initial Updates
This merges several sections into the new Development Fund Grants sections which outline the process for applying to the Dev Fund. The policy is also clarified to specify exactly how the fund can be used, this is important as we become a charity.

The limits on the Development Fund are increased, while the maximum amount of bids can be has increased in line with inflation since 2006, the bounds on the fund have stayed constant. We are thus applying a similar increase to the maximum and minimum bounds. With the increased size of SSAGO Events, we believe a loss of over £1,500 is not impossible in the future and the Dev Fund should be able to handle an extreme loss. 

The Dev Fund now has allowed use for loans to events, this has been done in the past although was never allowed in policy and is now explicitly given the power to be invested, which would allow a future exec the option to grow part of the fund.