Updates following Jack's Comments
Reformatting is introduced into section 4. to separate the different aspects of the grant.

Clarifications are provided around bids involving members of the national exec, (4.3.6-7.) the decision as to if the funds' aims are met remains at the trustee level to ensure spending is in line with these conditions. Reps is then given a final vote to approve the spending. An additional criteria is added for bids to explain their conflicts in the bid.

The amount low-on-funds is given explicitly and allows the exec to consider the current cash flow of SSAGO in differing decisions. In 2.5.